Add files via upload
This is a test app for the get_AspectRatio logic algorithm that I will use on a custom event emitter to adjust an array of class elements as the browser window changes.  This app version is an animated button that calls the app.  The button does not change dimensions, it was designed to display the ratio for testing purposes.  So far the logic is 100% accurate and requires a fraction of the memory used on previous sort and factorial concatenating algorithms.  The next step is to build the event emitter and test multiple elements...